Use pkg-config to locate -lprofiler and -ltcmalloc
There is no guarantee that these libraries are available on "standard"
locations, which means that simply linking to them with -l flags is
bound to fail.

For example: a user may have installed gperftools under their home
directory, in which case this crate cannot find it.

To resolve this, use pkg-config to locate the libraries.  This allows
this crate to learn all compiler and linker flags (in particular, -L)
needed to use them.